requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
flake8==4.0.1
flake8-isort==4.1.1
idna==3.3
isort==5.10.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyTelegramBotAPI==4.2.1
python-dotenv==0.19.2
requests==2.26.0
schedule==1.1.0
testfixtures==6.18.3
urllib3==1.26.7
